requirements/dev-requirements.txt
pytest
pytest-mock
pytest-cov
pep8
unittest2
astroid==1.4.0
pylint==1.5
